// src/modules/invoice/usecase/find-invoice/find-invoice.dto.ts
export interface FindInvoiceInputDTO {
  id: string;
}

export interface FindInvoiceOutputDTO {
  id: string;
  name: string;
  document: string;
  address: {
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
  };
  items: {
    id: string;
    name: string;
    price: number;
  }[];
  total: number;
  createdAt: Date;
}